src/ingestion-service/requirements.txt
confluent-kafka
faker
praw
requests
python-dotenv
atproto
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Pillow
